Improve generated files backup logic and regen autoconf files
In order for the Makefile.am changes in the debian patches to take
effect, we must regenerate the autoconf and automake files. This also
means a lot more generated files to back up.

Additionally, debian/rules (after backing up) the upstream
manpages (and all other generated files), to force them to be regenerated.